ログメッセージの修正とAPIモデルの追加
`PersonModelsApiAccess.cs` でカスタムフェイススプライトのログメッセージを修正しました。
`RedactionApiAccess.cs` では、usingディレクティブの整理とメソッドコメントの更新を行い、引数を `RedactVideoRequest` から `ApiRedactVideoRequestModel` に変更しました。
新たに `ApiFaceFilterModel.cs`、`ApiFaceRedactionModel.cs`、`ApiRedactVideoRequestModel.cs`、`ApiRedactVideoResponseModel.cs` のモデルクラスを追加し、APIのリクエストおよびレスポンスの構造を明確化しました。
