+++ VideoIndexerAccessCore/VideoIndexerClient/ApiModel/ApiFaceRedactionModel.cs
@@ -1,0 +1,8 @@
+﻿namespace VideoIndexerAccessCore.VideoIndexerClient.ApiModel
+{
+    public class ApiFaceRedactionModel
+    {
+        public string blurringKind { get; set; } = "HighBlur";
+        public ApiFaceFilterModel FilterModel { get; set; } = new ApiFaceFilterModel();
+    }
+}
