+++ VideoIndexerAccessCore/VideoIndexerClient/ApiModel/ApiRedactVideoRequestModel.cs
@@ -1,0 +1,7 @@
+﻿namespace VideoIndexerAccessCore.VideoIndexerClient.ApiModel
+{
+    public class ApiRedactVideoRequestModel
+    {
+        public ApiFaceRedactionModel? ApiFaces { get; set; }
+    }
+}
