+++ VideoIndexerAccessCore/VideoIndexerClient/ApiModel/ApiRedactVideoResponseModel.cs
@@ -1,0 +1,8 @@
+﻿namespace VideoIndexerAccessCore.VideoIndexerClient.ApiModel
+{
+    public class ApiRedactVideoResponseModel
+    {
+        public string? errorType { get; set; }
+        public string? message { get; set; }
+    }
+}
